gt-backend/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
greenlet==2.0.2
imutils==0.5.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.3
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pytesseract==0.3.10
pytest==7.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
